refactor(today): レストランパネルと useRestaurantState を切り出す
Issue #213 Ph-2c に沿い、店タブ・メニュー一覧の表示責務を RestaurantPanel / MenuItemList に、関連クライアント状態を useRestaurantState に集約する。挙動は変えない。

Made-with: Cursor
