+++ src/app/today/ui-constants.ts
@@ -1,0 +1,3 @@
+/** タブ右クリック・長押しメニュー等の固定幅・高さ（画面内に収める計算用） */
+export const TAB_CONTEXT_MENU_W = 176;
+export const TAB_CONTEXT_MENU_H = 44;
